refactor(container): relocate method inspector packages
